搭好雏形

+++ tensquare_user/src/main/java/com/tensquare/user/service/UserService.java
@@ -9,7 +9,9 @@ import javax.persistence.criteria.Expression;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
 import javax.persistence.criteria.Selection;
+import javax.servlet.http.HttpServletRequest;
 
+import io.jsonwebtoken.Claims;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -18,12 +20,15 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.redis.core.RedisTemplate;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import org.springframework.transaction.annotation.Transactional;
 import util.IdWorker;
 
 import com.tensquare.user.dao.UserDao;
 import com.tensquare.user.pojo.User;
+import util.JwtUtil;
 
 /**
  * 服务层
@@ -45,6 +50,15 @@ public class UserService {
 
 	@Autowired
 	private RabbitTemplate rabbitTemplate;
+
+	@Autowired
+	private BCryptPasswordEncoder encoder;
+
+	@Autowired
+	private HttpServletRequest request;
+
+	@Autowired
+	private JwtUtil jwtUtil;
 
 	/**
 	 * 查询全部列表
@@ -94,6 +108,8 @@ public class UserService {
 	 */
 	public void add(User user) {
 		user.setId( idWorker.nextId()+"" );
+		//密码加密
+		user.setPassword(encoder.encode(user.getPassword()));
 		user.setFanscount(0);//粉丝数
 		user.setFollowcount(0);//关注数
 		user.setOnline(0L);//在线时长
@@ -112,10 +128,14 @@ public class UserService {
 	}
 
 	/**
-	 * 删除
+	 * 删除 必须有admin角色才能删除
 	 * @param id
 	 */
 	public void deleteById(String id) {
+		String token = (String) request.getAttribute("claims_admin");
+		if (token==null || "".equals(token)){
+			throw new RuntimeException("权限不足！");
+		}
 		userDao.deleteById(id);
 	}
 
@@ -188,11 +208,25 @@ public class UserService {
 		Map<String,String> map = new HashMap<>();
 		map.put("mobile",mobile);
 		map.put("checkcode",checkcode);
-		rabbitTemplate.convertAndSend("sms",map);
+		//rabbitTemplate.convertAndSend("sms",map);
 		//控制台显示一份（方便测试）
 		System.out.println("验证码："+checkcode);
 
 
 	}
 
+
+	public User login(String mobile, String password) {
+		User user = userDao.findByMobile(mobile);
+		if (user!=null && encoder.matches(password,user.getPassword())){
+			return user;
+		}
+		return null;
+	}
+
+	@Transactional
+    public void updatefanscountandfollowcount(int x, String userid, String friendid) {
+		userDao.updatefanscount(x,friendid);
+		userDao.updatefollowcount(x,userid);
+    }
 }
